Refactoring sponza shadow test
