+++ Gem/Code/Source/ShadowedSponzaExampleComponent.h
