--- Scripts/ExpectedScreenshots/ShadowedBistro/directional_buildings.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:44b7b3178b006699c4b9d83699b458e46b44c6cd59a0b7ec4d9ea7a03e3ee39c
-size 633013
