--- Scripts/ExpectedScreenshots/ShadowedBistro/directional_filter_plaza.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:93b0875837d942f134ae206b005037a80f278a0360b95356ad7ff35dd5f78dd4
-size 370301
