--- Scripts/ExpectedScreenshots/ShadowedBistro/directional_filter_table.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:80af5cededa2567986272c9b0c9e6875efcc90ce9f6b9b183fdbe62d2ad8de74
-size 445192
