--- Scripts/ExpectedScreenshots/ShadowedBistro/directional_nofilter_plaza.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b215ce1fd77e26d78cf1a34524641351f4e60926c8bcae2a2beed75d5a397c39
-size 310296
