--- Scripts/ExpectedScreenshots/ShadowedBistro/directional_nofilter_table.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:95433fac5a35b6d3da16f91a6f1973694a668a54c3e88a3d8c5fb2690d7c6e82
-size 306545
