--- Scripts/ExpectedScreenshots/ShadowedBistro/initial.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:87e69755ca761ec3c96f74a592e5486a1eaa84f5398aab16f43e2e795ac064d1
-size 397841
