--- Scripts/ExpectedScreenshots/ShadowedBistro/spot_buildings.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d0f515167e69dc90a17fbb9dee77bbaf73ff7933564ad180cb31f893d408970e
-size 619386
