--- Scripts/ExpectedScreenshots/ShadowedBistro/spot_filter_table.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9010005e92cb7efef917c228bc43a69e80d59f8848250e0849bc88dbfa5779fe
-size 511459
