--- Scripts/ExpectedScreenshots/ShadowedBistro/spot_nofilter_table.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f8fb133d848712259a16c24c3f8fdbdf659db4dfc2cc8c7bef1886b1ebaf5806
-size 463588
