--- Scripts/ExpectedScreenshots/ShadowedSponza/directional_buildings.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a390a18302c1628b037b55cafa2f193569e54bdc822224b658ab60bde1b25f72
-size 559480
