+++ Scripts/ExpectedScreenshots/ShadowedSponza/directional_filter.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c4727ce981f94f07f513e24925699dcf8b099c80f6926c6f7484b30cdaf47fdc
+size 130413
