+++ Scripts/ExpectedScreenshots/ShadowedSponza/directional_nofilter.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b39ab344f23e50ba70984d38007a8f99f3ea81d8b1fcf438f082e95a19f611c8
+size 113493
