+++ Scripts/ExpectedScreenshots/ShadowedSponza/initial.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d22ae092cba0ef5f686d970e3feac105b86b852a73e2cb0f1608ed123d58e4fc
-size 141620
+oid sha256:674f6bd8229d6dc2892ff71a76239168648a1bdc9ee5768b4174aeb7211d1779
+size 141734
