--- Scripts/ExpectedScreenshots/ShadowedSponza/spot_buildings.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5290f20788b41e9721f78afd54a39db1b254d03b9e46764bbbf43e87281d0ff8
-size 492423
