+++ Scripts/ExpectedScreenshots/ShadowedSponza/spot_filter.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:cd608ea658492301d59e6bd241928e48e24183901c1963bd8771df72499aff5f
+size 146139
