--- Scripts/ExpectedScreenshots/ShadowedSponza/spot_filter_table.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9586f150ac2ec9d19f5fe84857bf6365bcb05081a6caf6e6445609cf4a792f9e
-size 9726
