+++ Scripts/ExpectedScreenshots/ShadowedSponza/spot_nofilter.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:78b7a109db6a4e42d863a83563c61d076278affd735a2efcfb3e18b3b245a6fa
+size 139876
